Integrate game_logic file into single file with rest of the game mechanics.
