+++ ocean/craftax_clean/constants.h
@@ -435,3 +435,25 @@ static const DungeonConfig DUNGEON_LEVEL_CONFIGS[3] = {
     {BLOCK_ENCHANTMENT_TABLE_ICE, BLOCK_WATER, BLOCK_WATER},
     {BLOCK_ENCHANTMENT_TABLE_FIRE, BLOCK_FOUNTAIN, BLOCK_PATH},
 };
+
+// Rendering parameters
+#define TEX_TILE_PX 16
+#define TEX_SCALE 3
+#define TEX_DRAW_PX (TEX_TILE_PX * TEX_SCALE)
+#define TEX_NUM (37 + 5 + 5 + 3 + 4)
+#define RENDER_ROWS 14
+#define RENDER_COLS 16
+#define ACTION_PANEL_W 280
+#define TEX_PLAYER_DOWN 37
+#define TEX_PLAYER_UP 38
+#define TEX_PLAYER_LEFT 39
+#define TEX_PLAYER_RIGHT 40
+#define TEX_PLAYER_SLEEP 41
+#define TEX_ITEM_BASE 42
+#define TEX_MOB_ZOMBIE 47
+#define TEX_MOB_SKELETON 48
+#define TEX_MOB_COW 49
+#define TEX_ARROW_DOWN 50
+#define TEX_ARROW_UP 51
+#define TEX_ARROW_LEFT 52
+#define TEX_ARROW_RIGHT 53
